+++ test/vietmap_plugin_test.dart
@@ -2,7 +2,22 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:vietmap_flutter_plugin/vietmap_flutter_plugin.dart';
 
 void main() {
+  Vietmap.getInstance("911f5b7e0a414ed6ab7e4a711c661fcc5a46fc9d1690e042");
   test('adds one to input values', () {
-    expect(Vietmap.autocomplete(VietMapAutoCompleteParams(textSearch: '')), []);
+    expectSync(Vietmap.autocomplete(VietMapAutoCompleteParams(textSearch: '')),
+        Future<Either<Failure, List<VietmapAutocompleteModel>>>);
+    expectSync(Vietmap.geoCode(VietMapAutoCompleteParams(textSearch: '')),
+        Future<Either<Failure, List<VietmapAutocompleteModel>>>);
+    expectSync(Vietmap.reverse(const LatLng(21.027763, 105.834160)),
+        Future<Either<Failure, VietmapReverseModel>>);
+    expectSync(Vietmap.place('Place ID'), []);
+    expectSync(Vietmap.getVietmapStyleUrl(),
+        'https://maps.vietmap.vn/api/maps/light/styles.json?apikey=911f5b7e0a414ed6ab7e4a711c661fcc5a46fc9d1690e042');
+    expectSync(
+        Vietmap.routing(VietMapRoutingParams(points: [
+          const LatLng(10.779391, 106.624833),
+          const LatLng(10.741039, 106.672116)
+        ])),
+        []);
   });
 }