+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-
+## 0.1.2, Jan 08, 2024
+* Add `pointsLatLng` property for `Paths` class, this property return list of `LatLng` of `Paths` object.
+## 0.1.1, Jan 05, 2024
+* Upgrade `vietmap_gl_platform_interface` to `0.1.5`
+* Export `VehicleType`, `LatLngExtension`, `StringExtension`, `VietmapModel`, `VietMapRouting`, `Paths`, `Instructions` class.
 ## 0.1.0, Dec 18, 2023
 * Upgrade `vietmap_gl_platform_interface` to `0.1.4`
 * Add topic for `pub.dev`
@@ -18,7 +22,6 @@
 * Update `README.md`
 * Downgrade min dart version support to `2.17.0`
 * Remove unused code
-
 ## 0.0.1, Nov 20, 2023
 * Initial release